Add assertRevert helper. Add createDataOrder helper.
