+++ test/TestDataOrder/TestDataOrderAddDataResponse.js
@@ -1,6 +1,6 @@
 const web3Utils = require('web3-utils');
 
-import { createHardcodedDataOrder } from "./helpers/dataOrderCreation";
+import { createDataOrder } from "./helpers/dataOrderCreation";
 import signMessage from "../helpers/signMessage";
 
 contract('DataOrder', (accounts) => {
@@ -15,13 +15,13 @@ contract('DataOrder', (accounts) => {
   const dataHash = "9eea36c42a56b62380d05f8430f3662e7720da6d5be3bdd1b20bb16e9d";
 
   let signature;
-  let emptyDataHashSignature ;
+  let emptyDataHashSignature;
   let orderWithoutNotary;
   let order;
 
   beforeEach('setup DataOrder for each test', async function () {
-    orderWithoutNotary = await createHardcodedDataOrder(owner, buyer);
-    order = await createHardcodedDataOrder(owner, buyer);
+    orderWithoutNotary = await createDataOrder({ buyer, from: owner });
+    order = await createDataOrder({ buyer, from: owner });
     await order.addNotary(
       notary,
       10,
