+++ test/TestDataOrder/TestDataOrderConstructor.js
@@ -1,33 +1,7 @@
 const DataOrder = artifacts.require("./DataOrder.sol");
 
-// TODO: Move to helpers
-const createDataOrder = async ({
-  buyer,
-  filters = "age:20,gender:male",
-  dataRequest = "data request",
-  price = 20,
-  initialBudgetForAudits = 10,
-  termsAndConditions = "DataOrder T&C",
-  buyerUrl = "https://buyer.example.com/data",
-  buyerPublicKey = "public-key",
-  owner
-}) => {
-  return await DataOrder.new(
-    buyer,
-    filters,
-    dataRequest,
-    price,
-    initialBudgetForAudits,
-    termsAndConditions,
-    buyerUrl,
-    buyerPublicKey,
-    { from: owner }
-  );
-}
-
-// TODO: Move to helpers
-const assertRevert = (error) =>
-  assert(error.toString().includes('revert'), error.toString());
+import { createDataOrder } from "./helpers/dataOrderCreation";
+import assertRevert from "../helpers/assertRevert";
 
 contract('DataOrder', async (accounts) => {
   const owner = accounts[0];
@@ -35,13 +9,13 @@ contract('DataOrder', async (accounts) => {
 
   describe('Constructor', async function () {
     it('creates a DataOrder', async function () {
-      const dataOrder = await createDataOrder({ owner, buyer });
+      const dataOrder = await createDataOrder({ buyer, from: owner });
       assert(dataOrder, "DataOrder was not created properly");
     });
 
     it('can not create a DataOrder with Zero Address as Buyer', async function () {
       try {
-        await createDataOrder({ owner, buyer: '0x0' });
+        await createDataOrder({ buyer: '0x0', from: owner });
         assert.fail();
       } catch (error) {
         assertRevert(error);
@@ -50,7 +24,7 @@ contract('DataOrder', async (accounts) => {
 
     it('can not create a DataOrder with an empty Buyer URL', async function () {
       try {
-        await createDataOrder({ owner, buyer, buyerUrl: '' });
+        await createDataOrder({ buyer, buyerUrl: '', from: owner });
         assert.fail();
       } catch (error) {
         assertRevert(error);
@@ -59,7 +33,7 @@ contract('DataOrder', async (accounts) => {
 
     it('can not create a DataOrder with an empty Buyer Public Key', async function () {
       try {
-        await createDataOrder({ owner, buyer, buyerPublicKey: '' });
+        await createDataOrder({ buyer, buyerPublicKey: '', from: owner });
         assert.fail();
       } catch (error) {
         assertRevert(error);
@@ -67,27 +41,27 @@ contract('DataOrder', async (accounts) => {
     });
 
     it('creates a DataOrder with the Sender as Buyer', async function () {
-      const dataOrder = await createDataOrder({ owner, buyer: owner });
+      const dataOrder = await createDataOrder({ buyer: owner, from: owner });
       assert(dataOrder, "The buyer can be the sender of the transaction");
     });
 
     it('creates a DataOrder with zero Price', async function () {
-      const dataOrder = await createDataOrder({ owner, buyer, price: 0 });
+      const dataOrder = await createDataOrder({ buyer, price: 0, from: owner });
       assert(dataOrder, "The price can be zero");
     });
 
     it('creates a DataOrder with empty Filters', async function () {
-      const dataOrder = await createDataOrder({ owner, buyer, filters: '' });
+      const dataOrder = await createDataOrder({ buyer, filters: '', from: owner });
       assert(dataOrder, "Filters can be empty");
     });
 
     it('creates a DataOrder with empty Data Request', async function () {
-      const dataOrder = await createDataOrder({ owner, buyer, dataRequest: '' });
+      const dataOrder = await createDataOrder({ buyer, dataRequest: '', from: owner });
       assert(dataOrder, "Data Request can be empty");
     });
 
     it('creates a DataOrder with empty Terms and Conditions', async function () {
-      const dataOrder = await createDataOrder({ owner, buyer, termsAndConditions: '' });
+      const dataOrder = await createDataOrder({ buyer, termsAndConditions: '', from: owner });
       assert(dataOrder, "Terms and Conditions can be empty");
     });
   });
